feat: import previously develop service
add base for service by declaring a SQS Resource, privileges to manage S3 resources, Document class
abstraction to handle flows for different types of documents (CreditReportDispute, GeneralDispute).
Bessides, this adds a new "engine" concept which is the one with the responsability to make the
parsing, PDFEngine take a template and using an instance of chromium it can create a PDF from a
compiled html with variables and a handlebars template.

"re https://github.com/debtcollective/disputes/issues/36"